fix(primitives)!: ScriptNOfK's threshold is the ledger's signed int64, not u32
The ledger CDDL has typed script_n_of_k's threshold signed since Shelley
(int32, note: "Shelley uses VInt for the threshold in script_n_of_k"),
widened to int64 in Allegra (note: "Allegra switched to int64 for
script_n_of_k thresholds"), through Conway. Checked directly against
IntersectMBO/cardano-ledger's shelley.cddl, allegra.cddl, mary.cddl,
alonzo.cddl, babbage.cddl and conway.cddl, and against the Haskell
RequireMOf/isValidMOf implementation (Int, i.e. 64-bit on every platform
cardano-node ships on). Pallas decoded and encoded it as u32 everywhere
(pallas-primitives' alonzo model, reused by babbage/conway; the separate
dijkstra model; and a second, independent NativeScript in pallas-network's
local-tx-submission miniprotocol).

A transaction with a negative threshold is valid, ordinary CBOR under the
real spec, and is trivially satisfied by any witness set (any non-negative
count is >= a negative number), which is why real cardano-node accepts it.
Pallas failed to decode it at all.

Reproduced against preprod tx
b1db2a411cb651a413840d3c8b112895a5bda2519a8ba6a372b8dd1ffc7746c2 (block
5184855, slot 133902219), which carries exactly this shape and stopped
Dolos cold: every restart resynced to the same point and hit the same
decode error, since the transaction is permanently on-chain.

Wire format is unaffected for every previously-representable value:
minicbor's signed-integer encoder delegates to the unsigned one for
non-negative inputs, producing byte-identical CBOR, and the signed
decoder is a strict superset of the unsigned one. No native-script hash,
policy ID or script address changes.

Also fixes the two places downstream of the primitive that assumed the
old type: phase1 native-script satisfaction in pallas-validate (now
compares in int64, matching isValidMOf's n <= 0 case) and the UtxoRPC
mapping, where the wire schema's ScriptNOfK.k is a stable uint32, so an
out-of-range threshold saturates on the way out instead of failing to
build (0 is the faithful image of any threshold <= 0, since both are
trivially satisfied; the upper clamp is unreachable in practice, since a
native script list is bounded by max tx size).

BREAKING CHANGE: NativeScript::ScriptNOfK's first field changes from u32
to i64 in pallas-primitives (alonzo, reused by babbage and conway; and
dijkstra) and in pallas-network's local-tx-submission NativeScript. Any
downstream code that pattern-matches or constructs this variant against
u32 will need updating.

+++ pallas-network/src/miniprotocols/localtxsubmission/primitives.rs
@@ -183,8 +183,10 @@ pub enum NativeScript {
     ScriptAll(Vec<NativeScript>),
     /// Require any sub-script to be satisfied.
     ScriptAny(Vec<NativeScript>),
-    /// Require at least `n` of the listed sub-scripts to be satisfied.
-    ScriptNOfK(u32, Vec<NativeScript>),
+    /// Require at least `n` of the listed sub-scripts to be satisfied. The
+    /// ledger CDDL types `n` signed (`int32` in Shelley, `int64` from
+    /// Allegra), not `uint`.
+    ScriptNOfK(i64, Vec<NativeScript>),
     /// Valid only at or after the given slot.
     InvalidBefore(u64),
     /// Valid only before the given slot.
